requirements.txt
# Python packages requirements
pyserial~=3.5          # Serial link to communicate with the display
PyYAML~=6.0.2          # For themes files
psutil~=7.0.0          # CPU / disk / network metrics
pystray~=0.19.5        # Tray icon (all OS)
babel~=2.17.0          # Date/time formatting
ruamel.yaml~=0.18.14    # For configuration editor
sv-ttk~=2.6.1          # Tk Sun Valley theme for configuration editor
tkinter-tooltip~=3.1.2 # Tooltips for configuration editor
uptime~=3.0.1          # For System Uptime
requests~=2.32.4       # HTTP library
ping3~=4.0.8           # ICMP ping implementation using raw socket
pyinstaller~=6.14.2    # bundles a Python application and all its dependencies into a single package
Pillow~=11.3.0         # Image generation

# Efficient image serialization
numpy~=2.0.2; python_version < "3.10"   # For Python 3.9, only numpy 2.0.x is supported
numpy~=2.2.6; python_version == "3.10"  # For Python 3.10, only numpy 2.2.x is supported
numpy~=2.3.1; python_version > "3.10"   # For Python > 3.10, any numpy 2.x is fine

# For Nvidia GPU on all platforms
GPUtil~=1.4.0; python_version < "3.12"
# GPUtil is broken for Python 3.12+ and not maintained anymore: fetch it from a fork where it is fixed
GPUtil @ git+https://github.com/mathoudebine/gputil.git@1.4.1-py3.13 ; python_version >= "3.12"

# Following packages are for AMD GPU on Linux
pyamdgpuinfo~=2.1.6; sys_platform=="linux" and python_version <= "3.12"
# For Python > 3.12, you need to build install pyamdgpuinfo manually: see https://github.com/mathoudebine/turing-smart-screen-python/wiki/Troubleshooting#linux--macos-no-supported-gpu-found-with-an-amd-gpu-and-python--311

# Following packages are for AMD GPU on Windows
pyadl~=0.1; sys_platform=="win32"

# Following packages are for Linux sleep/wake monitoring (D-Bus)
dbus-python~=1.4.0; sys_platform=="linux"
PyGObject~=3.56.1; sys_platform=="linux"

# Following packages are for LibreHardwareMonitor integration on Windows
pythonnet~=3.0.5; sys_platform=="win32"
pywin32>=306; sys_platform=="win32"
